stills an error

+++ src/main/java/dz/ibnrochd/master14/business/IConsultationService.java
@@ -17,5 +17,5 @@ public interface IConsultationService {
 	
 	void deleteConsultation(Consultation consultation);
 	
-	void updateConsultation(int id , Consultation consultation);
+	void updateConsultation(Long id , Consultation consultation);
 }
